fixing some more compiler warnings and some compiler errors submitted by Emanuel Rumpf

+++ rezound/src/PoolFile/TPoolFile.cpp
@@ -751,7 +751,7 @@ template<class l_addr_t,class p_addr_t>
 			return i->first;
 	}
 
-	printf("uh oh.. pool name not found for pool id: %u\n",poolId);
+	printf("uh oh.. pool name not found for pool id: %u\n",(unsigned)poolId);
 	exit(0);
 	return "";
 }
@@ -901,7 +901,7 @@ template<class l_addr_t,class p_addr_t>
 			if(logicalBlock.logicalStart!=expectedStart)
 			{
 				printSAT();
-				printf("pool: %u -- logical start wasn't what was expected in logical block: %lld\n",poolId,(long long)expectedStart);
+				printf("pool: %u -- logical start wasn't what was expected in logical block: %lld\n",(unsigned)poolId,(long long)expectedStart);
 				logicalBlock.print();
 				exit(1);
 			}
@@ -909,7 +909,7 @@ template<class l_addr_t,class p_addr_t>
 			if(!pasm.isAlloced(logicalBlock.physicalStart))
 			{
 				printSAT();
-				printf("pool: %u -- physicalStart isn't allocated\n",poolId);
+				printf("pool: %u -- physicalStart isn't allocated\n",(unsigned)poolId);
 				logicalBlock.print();
 				exit(1);
 			}
@@ -917,7 +917,7 @@ template<class l_addr_t,class p_addr_t>
 			if(pasm.getAllocedSize(logicalBlock.physicalStart)!=logicalBlock.size)
 			{
 				printSAT();
-				printf("pool: %u physical block's size isn't the same as the logical block's:\n",poolId);
+				printf("pool: %u physical block's size isn't the same as the logical block's:\n",(unsigned)poolId);
 				logicalBlock.print();
 				printf("physical block's size: %lld\n",(long long)pasm.getAllocedSize(logicalBlock.physicalStart));
 				exit(1);
@@ -930,7 +930,7 @@ template<class l_addr_t,class p_addr_t>
 				if((logicalBlock.physicalStart+logicalBlock.size)==nextLogicalBlock.physicalStart && 
 				   (logicalBlock.size+nextLogicalBlock.size)<=maxBlockSize)
 				{
-					printf("NOTE: two blocks could be joined: %u and %u\n",t,t+1);
+					printf("NOTE: two blocks could be joined: %u and %u\n",(unsigned)t,(unsigned)t+1);
 					logicalBlock.print();
 					nextLogicalBlock.print();
 				}
@@ -946,9 +946,9 @@ template<class l_addr_t,class p_addr_t>
 					if(CPhysicalAddressSpaceManager::overlap(logicalBlock.physicalStart,logicalBlock.size,SAT[x][y].physicalStart,SAT[x][y].size))
 					{
 						printSAT();
-						printf("pool: %u -- two blocks are occupying the same physical space\n",poolId);
+						printf("pool: %u -- two blocks are occupying the same physical space\n",(unsigned)poolId);
 						logicalBlock.print();
-						printf("and pool: %u\n",x);
+						printf("and pool: %u\n",(unsigned)x);
 						SAT[x][y].print();
 						exit(1);
 					}
@@ -962,7 +962,7 @@ template<class l_addr_t,class p_addr_t>
 			if((b.logicalStart+b.size)!=getPoolSize(poolId))
 			{
 				printSAT();
-				printf("pool: %u -- last address doesn't equal up to poolSize\n",poolId);
+				printf("pool: %u -- last address doesn't equal up to poolSize\n",(unsigned)poolId);
 				exit(1);
 			}
 
@@ -1436,10 +1436,10 @@ template<class l_addr_t,class p_addr_t>
 		if(!pools[poolId].isValid)
 			continue;
 
-		printf("\t%-4u Pool: '%s' size: %lld alignment: %lld\n",poolId,getPoolNameById(poolId).c_str(),(long long)getPoolSize(poolId),(long long)getPoolAlignment(poolId));
+		printf("\t%-4u Pool: '%s' size: %lld alignment: %lld\n",(unsigned)poolId,getPoolNameById(poolId).c_str(),(long long)getPoolSize(poolId),(long long)getPoolAlignment(poolId));
 		for(size_t t=0;t<SAT[poolId].size();t++)
 		{
-			printf("\t\t%-4u ",t);
+			printf("\t\t%-4u ",(unsigned)t);
 			SAT[poolId][t].print();
 		}
